refactor: split out root resolution and watchlist functions
