+++ root.c
@@ -6,10 +6,6 @@
 # include <sys/attr.h>
 #endif
 
-static w_ht_t *watched_roots = NULL;
-volatile long live_roots = 0;
-static pthread_mutex_t root_lock = PTHREAD_MUTEX_INITIALIZER;
-
 /* Ensure that we're synchronized with the state of the
  * filesystem at the current time.
  * We do this by touching a cookie file and waiting to
@@ -459,337 +455,6 @@ void w_root_teardown(w_root_t *root)
   }
 }
 
-static w_ht_val_t root_copy_val(w_ht_val_t val)
-{
-  w_root_t *root = w_ht_val_ptr(val);
-
-  w_root_addref(root);
-
-  return val;
-}
-
-static void root_del_val(w_ht_val_t val)
-{
-  w_root_t *root = w_ht_val_ptr(val);
-
-  w_root_delref(root);
-}
-
-static const struct watchman_hash_funcs root_funcs = {
-  w_ht_string_copy,
-  w_ht_string_del,
-  w_ht_string_equal,
-  w_ht_string_hash,
-  root_copy_val,
-  root_del_val
-};
-
-void watchman_watcher_init(void) {
-  watched_roots = w_ht_new(4, &root_funcs);
-}
-
-bool remove_root_from_watched(
-    w_root_t *root /* don't care about locked state */) {
-  bool removed = false;
-  pthread_mutex_lock(&root_lock);
-  // it's possible that the root has already been removed and replaced with
-  // another, so make sure we're removing the right object
-  if (w_ht_val_ptr(w_ht_get(watched_roots, w_ht_ptr_val(root->root_path))) ==
-      root) {
-    w_ht_del(watched_roots, w_ht_ptr_val(root->root_path));
-    removed = true;
-  }
-  pthread_mutex_unlock(&root_lock);
-  return removed;
-}
-
-/* Returns true if the global config root_restrict_files is not defined or if
- * one of the files in root_restrict_files exists, false otherwise. */
-static bool root_check_restrict(const char *watch_path)
-{
-  json_t *root_restrict_files = NULL;
-  uint32_t i;
-  bool enforcing;
-
-  root_restrict_files = cfg_compute_root_files(&enforcing);
-  if (!root_restrict_files) {
-    return true;
-  }
-  if (!enforcing) {
-    json_decref(root_restrict_files);
-    return true;
-  }
-
-  for (i = 0; i < json_array_size(root_restrict_files); i++) {
-    json_t *obj = json_array_get(root_restrict_files, i);
-    const char *restrict_file = json_string_value(obj);
-    char *restrict_path;
-    bool rv;
-
-    if (!restrict_file) {
-      w_log(W_LOG_ERR, "resolve_root: global config root_restrict_files "
-            "element %" PRIu32 " should be a string\n", i);
-      continue;
-    }
-
-    ignore_result(asprintf(&restrict_path, "%s%c%s", watch_path,
-          WATCHMAN_DIR_SEP, restrict_file));
-    rv = w_path_exists(restrict_path);
-    free(restrict_path);
-    if (rv)
-      return true;
-  }
-
-  return false;
-}
-
-static bool check_allowed_fs(const char *filename, char **errmsg)
-{
-  w_string_t *fs_type = w_fstype(filename);
-  json_t *illegal_fstypes = NULL;
-  json_t *advice_string;
-  uint32_t i;
-  const char *advice = NULL;
-
-  // Report this to the log always, as it is helpful in understanding
-  // problem reports
-  w_log(W_LOG_ERR, "path %s is on filesystem type %.*s\n",
-      filename, fs_type->len, fs_type->buf);
-
-  illegal_fstypes = cfg_get_json(NULL, "illegal_fstypes");
-  if (!illegal_fstypes) {
-    w_string_delref(fs_type);
-    return true;
-  }
-
-  advice_string = cfg_get_json(NULL, "illegal_fstypes_advice");
-  if (advice_string) {
-    advice = json_string_value(advice_string);
-  }
-  if (!advice) {
-    advice = "relocate the dir to an allowed filesystem type";
-  }
-
-  if (!json_is_array(illegal_fstypes)) {
-    w_log(W_LOG_ERR,
-          "resolve_root: global config illegal_fstypes is not an array\n");
-    w_string_delref(fs_type);
-    return true;
-  }
-
-  for (i = 0; i < json_array_size(illegal_fstypes); i++) {
-    json_t *obj = json_array_get(illegal_fstypes, i);
-    const char *name = json_string_value(obj);
-
-    if (!name) {
-      w_log(W_LOG_ERR, "resolve_root: global config illegal_fstypes "
-            "element %" PRIu32 " should be a string\n", i);
-      continue;
-    }
-
-    if (!w_string_equal_cstring(fs_type, name)) {
-      continue;
-    }
-
-    ignore_result(asprintf(errmsg,
-      "path uses the \"%.*s\" filesystem "
-      "and is disallowed by global config illegal_fstypes: %s",
-      fs_type->len, fs_type->buf, advice));
-
-    w_string_delref(fs_type);
-    return false;
-  }
-
-  w_string_delref(fs_type);
-  return true;
-}
-
-// Given a filename, walk the current set of watches.
-// If a watch is a prefix match for filename then we consider it to
-// be an enclosing watch and we'll return the root path and the relative
-// path to filename.
-// Returns NULL if there were no matches.
-// If multiple watches have the same prefix, it is undefined which one will
-// match.
-char *w_find_enclosing_root(const char *filename, char **relpath) {
-  w_ht_iter_t i;
-  w_root_t *root = NULL;
-  w_string_t *name = w_string_new_typed(filename, W_STRING_BYTE);
-  char *prefix = NULL;
-
-  pthread_mutex_lock(&root_lock);
-  if (w_ht_first(watched_roots, &i)) do {
-    w_string_t *root_name = w_ht_val_ptr(i.key);
-    if (w_string_startswith(name, root_name) && (
-          name->len == root_name->len /* exact match */ ||
-          is_slash(name->buf[root_name->len]) /* dir container matches */)) {
-      root = w_ht_val_ptr(i.value);
-      w_root_addref(root);
-      break;
-    }
-  } while (w_ht_next(watched_roots, &i));
-  pthread_mutex_unlock(&root_lock);
-
-  if (!root) {
-    goto out;
-  }
-
-  // extract the path portions
-  prefix = malloc(root->root_path->len + 1);
-  if (!prefix) {
-    goto out;
-  }
-  memcpy(prefix, filename, root->root_path->len);
-  prefix[root->root_path->len] = '\0';
-
-  if (root->root_path->len == name->len) {
-    *relpath = NULL;
-  } else {
-    *relpath = strdup(filename + root->root_path->len + 1);
-  }
-
-out:
-  if (root) {
-    w_root_delref(root);
-  }
-  w_string_delref(name);
-
-  return prefix;
-}
-
-static bool root_resolve(const char *filename, bool auto_watch, bool *created,
-                         char **errmsg,
-                         struct unlocked_watchman_root *unlocked) {
-  struct watchman_root *root = NULL, *existing = NULL;
-  w_ht_val_t root_val;
-  char *watch_path;
-  w_string_t *root_str;
-  int realpath_err;
-
-  *created = false;
-  unlocked->root = NULL;
-
-  // Sanity check that the path is absolute
-  if (!w_is_path_absolute_cstr(filename)) {
-    ignore_result(asprintf(errmsg, "path \"%s\" must be absolute", filename));
-    w_log(W_LOG_ERR, "resolve_root: %s", *errmsg);
-    return false;
-  }
-
-  if (!strcmp(filename, "/")) {
-    ignore_result(asprintf(errmsg, "cannot watch \"/\""));
-    w_log(W_LOG_ERR, "resolve_root: %s", *errmsg);
-    return false;
-  }
-
-  watch_path = w_realpath(filename);
-  realpath_err = errno;
-
-  if (!watch_path) {
-    watch_path = (char*)filename;
-  }
-
-  root_str = w_string_new_typed(watch_path, W_STRING_BYTE);
-  pthread_mutex_lock(&root_lock);
-  // This will addref if it returns root
-  if (w_ht_lookup(watched_roots, w_ht_ptr_val(root_str), &root_val, true)) {
-    root = w_ht_val_ptr(root_val);
-  }
-  pthread_mutex_unlock(&root_lock);
-  w_string_delref(root_str);
-
-  if (!root && watch_path == filename) {
-    // Path didn't resolve and neither did the name they passed in
-    ignore_result(asprintf(errmsg,
-          "realpath(%s) -> %s", filename, strerror(realpath_err)));
-    w_log(W_LOG_ERR, "resolve_root: %s\n", *errmsg);
-    return false;
-  }
-
-  if (root || !auto_watch) {
-    if (!root) {
-      ignore_result(
-          asprintf(errmsg, "directory %s is not watched", watch_path));
-      w_log(W_LOG_DBG, "resolve_root: %s\n", *errmsg);
-    }
-    if (watch_path != filename) {
-      free(watch_path);
-    }
-
-    if (!root) {
-      return false;
-    }
-
-    // Treat this as new activity for aging purposes; this roughly maps
-    // to a client querying something about the root and should extend
-    // the lifetime of the root
-
-    unlocked->root = root;
-    // Note that this write potentially races with the read in consider_reap
-    // but we're "OK" with it because the latter is performed under a write
-    // lock and the worst case side effect is that we (safely) decide to reap
-    // at the same instant that a new command comes in.  The reap intervals
-    // are typically on the order of days.
-    time(&unlocked->root->last_cmd_timestamp);
-    // caller owns a ref
-    return true;
-  }
-
-  w_log(W_LOG_DBG, "Want to watch %s -> %s\n", filename, watch_path);
-
-  if (!check_allowed_fs(watch_path, errmsg)) {
-    w_log(W_LOG_ERR, "resolve_root: %s\n", *errmsg);
-    if (watch_path != filename) {
-      free(watch_path);
-    }
-    return false;
-  }
-
-  if (!root_check_restrict(watch_path)) {
-    ignore_result(
-        asprintf(errmsg, "Your watchman administrator has configured watchman "
-                         "to prevent watching this path.  None of the files "
-                         "listed in global config root_files are "
-                         "present and enforce_root_files is set to true"));
-    w_log(W_LOG_ERR, "resolve_root: %s\n", *errmsg);
-    if (watch_path != filename) {
-      free(watch_path);
-    }
-    return false;
-  }
-
-  // created with 1 ref
-  root = w_root_new(watch_path, errmsg);
-
-  if (watch_path != filename) {
-    free(watch_path);
-  }
-
-  if (!root) {
-    return false;
-  }
-
-  pthread_mutex_lock(&root_lock);
-  existing = w_ht_val_ptr(w_ht_get(watched_roots,
-                w_ht_ptr_val(root->root_path)));
-  if (existing) {
-    // Someone beat us in this race
-    w_root_addref(existing);
-    w_root_delref(root);
-    root = existing;
-    *created = false;
-  } else {
-    // adds 1 ref
-    w_ht_set(watched_roots, w_ht_ptr_val(root->root_path), w_ht_ptr_val(root));
-    *created = true;
-  }
-  pthread_mutex_unlock(&root_lock);
-
-  // caller owns 1 ref
-  unlocked->root = root;
-  return true;
-}
-
 static bool start_detached_root_thread(w_root_t *root, char **errmsg,
     void*(*func)(void*), pthread_t *thr) {
   pthread_attr_t attr;
@@ -812,7 +477,7 @@ static bool start_detached_root_thread(w_root_t *root, char **errmsg,
   return false;
 }
 
-static bool root_start(w_root_t *root, char **errmsg)
+bool root_start(w_root_t *root, char **errmsg)
 {
   if (!start_detached_root_thread(root, errmsg,
         run_notify_thread, &root->notify_thread)) {
@@ -831,41 +496,7 @@ static bool root_start(w_root_t *root, char **errmsg)
   return true;
 }
 
-bool w_root_resolve_for_client_mode(const char *filename, char **errmsg,
-                                    struct unlocked_watchman_root *unlocked) {
-  bool created = false;
-
-  if (!root_resolve(filename, true, &created, errmsg, unlocked)) {
-    return false;
-  }
-
-  if (created) {
-    struct timeval start;
-    struct watchman_pending_collection pending;
-    struct write_locked_watchman_root lock;
-
-    w_pending_coll_init(&pending);
-
-    /* force a walk now */
-    gettimeofday(&start, NULL);
-    w_root_lock(unlocked, "w_root_resolve_for_client_mode", &lock);
-    w_pending_coll_add(&lock.root->pending, lock.root->root_path,
-        start, W_PENDING_RECURSIVE);
-    while (w_root_process_pending(&lock, &pending, true)) {
-      // Note that we don't need a two-level loop (as we do in the main
-      // watcher-enabled mode) in client mode as we are not using a
-      // watcher in this situation.
-      ;
-    }
-    w_root_unlock(&lock, unlocked);
-
-    w_pending_coll_destroy(&pending);
-  }
-  return true;
-}
-
-static void signal_root_threads(w_root_t *root)
-{
+void signal_root_threads(w_root_t *root) {
   // Send SIGUSR1 to interrupt blocking syscalls on the
   // worker threads.  They'll self-terminate.
   if (!pthread_equal(root->notify_thread, pthread_self())) {
@@ -924,59 +555,6 @@ bool w_root_stop_watch(struct unlocked_watchman_root *unlocked)
   return stopped;
 }
 
-json_t *w_root_stop_watch_all(void)
-{
-  uint32_t roots_count, i;
-  w_root_t **roots;
-  w_ht_iter_t iter;
-  json_t *stopped;
-
-  pthread_mutex_lock(&root_lock);
-  roots_count = w_ht_size(watched_roots);
-  roots = calloc(roots_count, sizeof(*roots));
-
-  i = 0;
-  if (w_ht_first(watched_roots, &iter)) do {
-    w_root_t *root = w_ht_val_ptr(iter.value);
-    w_root_addref(root);
-    roots[i++] = root;
-  } while (w_ht_next(watched_roots, &iter));
-
-  stopped = json_array();
-  for (i = 0; i < roots_count; i++) {
-    w_root_t *root = roots[i];
-    w_string_t *path = root->root_path;
-    if (w_ht_del(watched_roots, w_ht_ptr_val(path))) {
-      w_root_cancel(root);
-      json_array_append_new(stopped, w_string_to_json(path));
-    }
-    w_root_delref(root);
-  }
-  free(roots);
-  pthread_mutex_unlock(&root_lock);
-
-  w_state_save();
-
-  return stopped;
-}
-
-bool w_root_resolve(const char *filename, bool auto_watch, char **errmsg,
-                    struct unlocked_watchman_root *unlocked) {
-  bool created = false;
-  if (!root_resolve(filename, auto_watch, &created, errmsg, unlocked)) {
-    return false;
-  }
-  if (created) {
-    if (!root_start(unlocked->root, errmsg)) {
-      w_root_cancel(unlocked->root);
-      w_root_delref(unlocked->root);
-      return false;
-    }
-    w_state_save();
-  }
-  return true;
-}
-
 // Caller must have locked root
 json_t *w_root_trigger_list_to_json(struct read_locked_watchman_root *lock)
 {
@@ -991,177 +569,6 @@ json_t *w_root_trigger_list_to_json(struct read_locked_watchman_root *lock)
   } while (w_ht_next(lock->root->commands, &iter));
 
   return arr;
-}
-
-json_t *w_root_watch_list_to_json(void)
-{
-  w_ht_iter_t iter;
-  json_t *arr;
-
-  arr = json_array();
-
-  pthread_mutex_lock(&root_lock);
-  if (w_ht_first(watched_roots, &iter)) do {
-    w_root_t *root = w_ht_val_ptr(iter.value);
-    json_array_append_new(arr, w_string_to_json(root->root_path));
-  } while (w_ht_next(watched_roots, &iter));
-  pthread_mutex_unlock(&root_lock);
-
-  return arr;
-}
-
-bool w_root_load_state(json_t *state)
-{
-  json_t *watched;
-  size_t i;
-
-  watched = json_object_get(state, "watched");
-  if (!watched) {
-    return true;
-  }
-
-  if (!json_is_array(watched)) {
-    return false;
-  }
-
-  for (i = 0; i < json_array_size(watched); i++) {
-    json_t *obj = json_array_get(watched, i);
-    bool created = false;
-    const char *filename;
-    json_t *triggers;
-    size_t j;
-    char *errmsg = NULL;
-    struct write_locked_watchman_root lock;
-    struct unlocked_watchman_root unlocked;
-
-    triggers = json_object_get(obj, "triggers");
-    filename = json_string_value(json_object_get(obj, "path"));
-    if (!root_resolve(filename, true, &created, &errmsg, &unlocked)) {
-      free(errmsg);
-      continue;
-    }
-
-    w_root_lock(&unlocked, "w_root_load_state", &lock);
-
-    /* re-create the trigger configuration */
-    for (j = 0; j < json_array_size(triggers); j++) {
-      json_t *tobj = json_array_get(triggers, j);
-      json_t *rarray;
-      struct watchman_trigger_command *cmd;
-
-      // Legacy rules format
-      rarray = json_object_get(tobj, "rules");
-      if (rarray) {
-        continue;
-      }
-
-      cmd = w_build_trigger_from_def(lock.root, tobj, &errmsg);
-      if (!cmd) {
-        w_log(W_LOG_ERR, "loading trigger for %s: %s\n",
-              lock.root->root_path->buf, errmsg);
-        free(errmsg);
-        continue;
-      }
-
-      w_ht_replace(lock.root->commands, w_ht_ptr_val(cmd->triggername),
-                   w_ht_ptr_val(cmd));
-    }
-    w_root_unlock(&lock, &unlocked);
-
-    if (created) {
-      if (!root_start(unlocked.root, &errmsg)) {
-        w_log(W_LOG_ERR, "root_start(%s) failed: %s\n",
-            unlocked.root->root_path->buf, errmsg);
-        free(errmsg);
-        w_root_cancel(unlocked.root);
-      }
-    }
-
-    w_root_delref(unlocked.root);
-  }
-
-  return true;
-}
-
-bool w_root_save_state(json_t *state)
-{
-  w_ht_iter_t root_iter;
-  bool result = true;
-  json_t *watched_dirs;
-
-  watched_dirs = json_array();
-
-  w_log(W_LOG_DBG, "saving state\n");
-
-  pthread_mutex_lock(&root_lock);
-  if (w_ht_first(watched_roots, &root_iter)) do {
-    json_t *obj;
-    json_t *triggers;
-    struct read_locked_watchman_root lock;
-    struct unlocked_watchman_root unlocked = {w_ht_val_ptr(root_iter.value)};
-
-    obj = json_object();
-
-    json_object_set_new(obj, "path",
-                        w_string_to_json(unlocked.root->root_path));
-
-    w_root_read_lock(&unlocked, "w_root_save_state", &lock);
-    triggers = w_root_trigger_list_to_json(&lock);
-    w_root_read_unlock(&lock, &unlocked);
-    json_object_set_new(obj, "triggers", triggers);
-
-    json_array_append_new(watched_dirs, obj);
-
-  } while (w_ht_next(watched_roots, &root_iter));
-
-  pthread_mutex_unlock(&root_lock);
-
-  json_object_set_new(state, "watched", watched_dirs);
-
-  return result;
-}
-
-void w_root_free_watched_roots(void)
-{
-  w_ht_iter_t root_iter;
-  int last, interval;
-  time_t started;
-
-  // Reap any children so that we can release their
-  // references on the root
-  w_reap_children(true);
-
-  pthread_mutex_lock(&root_lock);
-  if (w_ht_first(watched_roots, &root_iter)) do {
-    w_root_t *root = w_ht_val_ptr(root_iter.value);
-    if (!w_root_cancel(root)) {
-      signal_root_threads(root);
-    }
-  } while (w_ht_next(watched_roots, &root_iter));
-  pthread_mutex_unlock(&root_lock);
-
-  last = live_roots;
-  time(&started);
-  w_log(W_LOG_DBG, "waiting for roots to cancel and go away %d\n", last);
-  interval = 100;
-  for (;;) {
-    int current = __sync_fetch_and_add(&live_roots, 0);
-    if (current == 0) {
-      break;
-    }
-    if (time(NULL) > started + 3) {
-      w_log(W_LOG_ERR, "%d roots were still live at exit\n", current);
-      break;
-    }
-    if (current != last) {
-      w_log(W_LOG_DBG, "waiting: %d live\n", current);
-      last = current;
-    }
-    usleep(interval);
-    interval = MIN(interval * 2, 1000000);
-  }
-
-  w_log(W_LOG_DBG, "all roots are gone\n");
 }
 
 /* vim:ts=2:sw=2:et:
